requirements.txt
distlib==0.3.1
Django==2.2.13
djangorestframework==3.9.4
virtualenv==20.4.2
